Break down position calcultions
