+++ packages/elements/src/components/tooltip/tooltip.tsx
@@ -1,37 +1,6 @@
 import React, { RefObject } from "react";
-import ReactDOM from "react-dom";
 import { TooltipContainer } from "./tooltip.styles";
-
-/**
- * A component that creates a tooltip portal container within which
- * tooltip components can be rendered. The tooltip portal is appended and removed
- * to/from the bottom of the document as needed.
- */
-export class TooltipPortal extends React.PureComponent {
-
-  /**
-   * The element to append to the document.
-   */
-  private element: HTMLDivElement;
-
-  constructor(props: any) {
-    super(props);
-    this.element = document.createElement("div");
-    this.element.classList.add("tooltip-portal");
-  }
-
-  componentDidMount(): void {
-    document.body.appendChild(this.element);
-  }
-
-  componentWillUnmount(): void {
-    document.body.removeChild(this.element);
-  }
-
-  render() {
-    return ReactDOM.createPortal(this.props.children, this.element);
-  }
-}
+import { TooltipPortal } from "./tooltip-portal";
 
 /**
  * Tooltip props.
@@ -62,35 +31,62 @@ export interface TooltipState {
 }
 
 export class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
-  private gap: number;
+  /**
+   * The gap to maintain between the tooltip, screen edges and the target element, in pixels.
+   */
+  private gap: number = 8
+
+  /**
+   * The target element to attach the tooltip to.
+   */
   private targetElement: HTMLSpanElement | null;
+
+  /**
+   * The tooltip element.
+   */
   private tooltipElement: RefObject<HTMLDivElement>;
+
+  /**
+   * Available tooltip positions.
+   */
   private availablePositions = ["top", "bottom", "left", "right"];
+
+  /**
+   * Default position to show tooltip in.
+   */
   private defaultPosition = "bottom";
+
+  /**
+   * Default class for tooltip.
+   */
   private defaultClass: string = "tooltip";
 
   constructor(props: TooltipProps) {
     super(props);
     this.targetElement = null;
     this.tooltipElement = React.createRef<HTMLDivElement>();
-
     this.state = {
       visible: false,
     };
-
-    this.gap = 8;
   }
 
+  /**
+   * Shows the tooltip.
+   */
   showTooltip() {
-    this.setState({
-      visible: true,
-    });
+    this.setState({ visible: true });
   }
 
+  /**
+   * Hides the tooltip.
+   */
   hideTooltip() {
     this.setState({ visible: false });
   }
 
+  /**
+   * Returns a string of applicable position classes.
+   */
   getPositionClass(): string {
     if (
       this.props.position &&
@@ -102,21 +98,27 @@ export class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
     }
   }
 
+  /**
+   * Returns a string of all applicable classes.
+   */
   getClasses(): string {
     return [this.defaultClass, this.getPositionClass()].join(" ").trim();
   }
 
+  /**
+   * Wait for the tooltip to be added to the DOM and update it's position,
+   * and then display it.
+   */
   componentDidUpdate(): void {
     this.updateTooltipPosition();
     this.displayTooltip();
   }
 
-  displayTooltip(): void {
-    if (!!!this.tooltipElement.current) {
-      return;
-    }
-    this.tooltipElement.current.style.visibility = "visible";
-  }
+  /**
+   * Update the position of the tooltip on the screen based on the current
+   * position prop, tooltip dimensions and the position and dimensions
+   * of the target element.
+   */
 
   updateTooltipPosition(): void {
     if (!!!this.targetElement || !!!this.tooltipElement.current) {
@@ -126,67 +128,95 @@ export class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
     const targetDimensions = this.targetElement.getBoundingClientRect();
     const tooltipDimensions = this.tooltipElement.current?.getBoundingClientRect();
 
-    switch (this.props.position) {
-      case "top":
-        this.tooltipElement.current.style.bottom =
-          window.innerHeight - targetDimensions.top + this.gap + "px";
-        this.tooltipElement.current.style.left = this.getVerticalLeft(
-          targetDimensions.left,
-          targetDimensions.width,
-          tooltipDimensions.width
-        );
-        break;
+    this.tooltipElement.current.style.left = this.getLeft(
+      targetDimensions.left,
+      targetDimensions.width,
+      tooltipDimensions.width
+    );
 
-      case "right":
-        this.tooltipElement.current.style.top = this.getHorizontalTop(
-          targetDimensions.top,
-          targetDimensions.height,
-          tooltipDimensions.height
-        );
-        this.tooltipElement.current.style.left =
-          targetDimensions.left + targetDimensions.width + this.gap + "px";
-        break;
-
-      case "left":
-        this.tooltipElement.current.style.top = this.getHorizontalTop(
-          targetDimensions.top,
-          targetDimensions.height,
-          tooltipDimensions.height
-        );
-        this.tooltipElement.current.style.left =
-          targetDimensions.left - tooltipDimensions.width - this.gap * 2 + "px";
-        break;
-
-      case "bottom":
-      default:
-        this.tooltipElement.current.style.top =
-          targetDimensions.top + targetDimensions.height + this.gap + "px";
-        this.tooltipElement.current.style.left = this.getVerticalLeft(
-          targetDimensions.left,
-          targetDimensions.width,
-          tooltipDimensions.width
-        );
-        break;
-    }
+    this.tooltipElement.current.style.top = this.getTop(
+      targetDimensions.top,
+      targetDimensions.height,
+      tooltipDimensions.height
+    );
   }
 
-  getVerticalLeft(
+  /**
+   * Set the tooltip visibility to visible.
+   */
+  displayTooltip(): void {
+    if (!!!this.tooltipElement.current) {
+      return;
+    }
+    this.tooltipElement.current.style.visibility = "visible";
+  }
+
+  /**
+   * Returns the left value in pixels for the tooltip based on the
+   * current position prop.
+   * @param targetLeft the left offset for the target element
+   * @param targetWidth the width of the target element
+   * @param tooltipWidth the width of the tooltip
+   */
+  getLeft(
     targetLeft: number,
     targetWidth: number,
     tooltipWidth: number
   ): string {
-    let left = targetLeft + targetWidth / 2 - tooltipWidth / 2;
-    left = Math.max(this.gap, left);
-    left = Math.min(left, document.body.clientWidth - tooltipWidth + this.gap);
+    let left: number = 0;
+
+    switch (this.props.position) {
+      case "right":
+        left = targetLeft + targetWidth + this.gap;
+        break;
+
+      case "left":
+        left = targetLeft - tooltipWidth - this.gap * 2;
+        break;
+
+      case "top":
+      case "bottom":
+      default:
+        left = targetLeft + targetWidth / 2 - tooltipWidth / 2;
+        left = Math.max(this.gap, left);
+        left = Math.min(left, document.body.clientWidth - tooltipWidth + this.gap);
+        break;
+    }
+
     return left + "px";
   }
 
-  getHorizontalTop(
+  /**
+   * Returns the top value in pixels for the tooltip based on the
+   * current position prop.
+   * @param targetTop the top offset for the target element
+   * @param targetHeight the height of the target element
+   * @param tooltipHeight the height of the tooltip
+   */
+  getTop(
     targetTop: number,
     targetHeight: number,
     tooltipHeight: number
   ): string {
-    return targetTop + targetHeight / 2 - tooltipHeight / 2 + "px";
+    let top: number = 0;
+
+    switch (this.props.position) {
+      case "top":
+        top = targetTop - tooltipHeight - this.gap;
+        break;
+
+      case "right":
+      case "left":
+        top = targetTop + targetHeight / 2 - tooltipHeight / 2;
+        break;
+
+      case "bottom":
+      default:
+        top = targetTop + targetHeight + this.gap;
+        break;
+    }
+
+    return top + "px";
   }
 
   render() {
